refactor: change route parameters from path params to query strings - Change optional parameters from path params (/:params) to query strings (?limit=N) - Remove unnecessary cache - Extract ranking-related types and constants to separate files - Extract isekai ranking handling to a separate file - Enhance code structure and readability
